Fix codes duplication
